requirements.txt
# Minimal runtime requirements discovered from imports in the repo
pandas>=1.5
numpy>=1.24
PyYAML>=6.0
rapidfuzz>=2.14.0
sentence-transformers>=2.2.2
scikit-learn>=1.2
prophet>=1.1
google-generativeai>=0.3.0
matplotlib>=3.7
reportlab>=4.0.4
markdown>=3.4

# Optional / dev (uncomment if used)
# jupyter
pytest>=7.4.0  # Required for running tests